use mem_factory in all concurr types
